[Backend] Điều chỉnh model Category

+++ src/graphql/modules/category/category.resolver.ts
@@ -1,3 +1,4 @@
+import { set } from "lodash";
 import { ErrorHelper } from "../../../base/error";
 import { ROLES } from "../../../constants/role.const";
 import { AuthHelper } from "../../../helpers";
@@ -8,54 +9,40 @@ import { categoryService } from "./category.service";
 
 const Query = {
   getAllCategory: async (root: any, args: any, context: Context) => {
-    AuthHelper.acceptRoles(context, ROLES.ADMIN_EDITOR_MEMBER_CUSTOMER);
+    context.auth(ROLES.ANONYMOUS_CUSTOMER_MEMBER);
+    set(args, "q.filter.memberId", context.sellerId);
     return categoryService.fetch(args.q);
   },
   getOneCategory: async (root: any, args: any, context: Context) => {
-    AuthHelper.acceptRoles(context, ROLES.ADMIN_EDITOR_MEMBER_CUSTOMER);
+    context.auth(ROLES.ANONYMOUS_CUSTOMER_MEMBER);
     const { id } = args;
     return await categoryService.findOne({ _id: id });
   },
   getFilteringCategories: async (root: any, args: any, context: Context) => {
-    return await CategoryModel.find({isPrimary:true});
+    return await CategoryModel.find({ isPrimary: true });
   },
 };
 
 const Mutation = {
   createCategory: async (root: any, args: any, context: Context) => {
-    AuthHelper.acceptRoles(context, ROLES.ADMIN_EDITOR_MEMBER);
+    context.auth([ROLES.MEMBER]);
     const { data } = args;
-    // member tao duoc san pham mo rong
-    if (context.tokenData.role == ROLES.MEMBER) {
-      data.memberId = context.tokenData._id;
-    } else {
-      // admin tao san pham chinh
-      data.isPrimary = true;
-    }
+    data.memberId = context.sellerId;
     data.code = data.code || (await CategoryHelper.generateCode());
     return await categoryService.create(data);
   },
   updateCategory: async (root: any, args: any, context: Context) => {
-    AuthHelper.acceptRoles(context, ROLES.ADMIN_EDITOR_MEMBER);
+    context.auth([ROLES.MEMBER]);
     const { id, data } = args;
-    if (context.tokenData.role == ROLES.MEMBER) {
-      const category = await CategoryModel.findById(id);
-      if (category.memberId.toString() != context.tokenData._id) {
-        throw ErrorHelper.permissionDeny();
-      }
-    }
+    await protectDoc(id, context);
     return await categoryService.updateOne(id, data);
   },
   deleteOneCategory: async (root: any, args: any, context: Context) => {
-    AuthHelper.acceptRoles(context, ROLES.ADMIN_EDITOR_MEMBER);
+    context.auth([ROLES.MEMBER]);
     const { id } = args;
+    const category = await protectDoc(id, context);
+    if (category.productIds.length > 0) throw Error("Không thể xoá khi còn sản phẩm");
     return await categoryService.deleteOne(id);
-  },
-  deleteManyCategory: async (root: any, args: any, context: Context) => {
-    AuthHelper.acceptRoles(context, [ROLES.ADMIN]);
-    const { ids } = args;
-    let result = await categoryService.deleteMany(ids);
-    return result;
   },
 };
 
@@ -66,3 +53,8 @@ export default {
   Mutation,
   Category,
 };
+async function protectDoc(id: any, context: Context) {
+  const category = await CategoryModel.findById(id);
+  if (category.memberId.toString() != context.sellerId) throw ErrorHelper.permissionDeny();
+  return category;
+}
